+++ src/Symfony/Component/BrowserKit/Response.php
@@ -13,11 +13,16 @@ namespace Symfony\Component\BrowserKit;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since Symfony 4.3
  */
 class Response
 {
+    /** @internal */
     protected $content;
+    /** @internal */
     protected $status;
+    /** @internal */
     protected $headers;
 
     /**
@@ -45,10 +50,10 @@ class Response
         $headers = '';
         foreach ($this->headers as $name => $value) {
             if (\is_string($value)) {
-                $headers .= $this->buildHeader($name, $value);
+                $headers .= sprintf("%s: %s\n", $name, $value);
             } else {
                 foreach ($value as $headerValue) {
-                    $headers .= $this->buildHeader($name, $headerValue);
+                    $headers .= sprintf("%s: %s\n", $name, $headerValue);
                 }
             }
         }
@@ -63,9 +68,13 @@ class Response
      * @param string $value The header value
      *
      * @return string The built header line
+     *
+     * @deprecated since Symfony 4.3
      */
     protected function buildHeader($name, $value)
     {
+        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.3.', __METHOD__), E_USER_DEPRECATED);
+
         return sprintf("%s: %s\n", $name, $value);
     }
 
@@ -83,8 +92,17 @@ class Response
      * Gets the response status code.
      *
      * @return int The response status code
+     *
+     * @deprecated since Symfony 4.3, use getStatusCode() instead
      */
     public function getStatus()
+    {
+        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.3, use getStatusCode() instead.', __METHOD__), E_USER_DEPRECATED);
+
+        return $this->status;
+    }
+
+    public function getStatusCode(): int
     {
         return $this->status;
     }